add: organized parsing by files
