+++ Makefile
@@ -13,7 +13,7 @@ vpath %.c $(SRC_DIRS)
 
 MAIN_FILE := main.c
 
-SRC_PARSING_FILES := parsing.c
+SRC_PARSING_FILES := parsing.c parsing_utils.c parse_metadata.c parse_map.c
 SRC_PARSING := $(addprefix parsing/, $(SRC_PARSING_FILES))
 
 SRC_INIT_FILES := initialize.c
