+++ map.ber
@@ -35,4 +35,10 @@ C 123,121,0
         1001
         1001
         1001
+        1001
+        1001
+        1001
+        1001
+        1001
+        1001
         1111
