+++ src/raycasting/draw_rays.c
@@ -6,7 +6,7 @@
 /*   By: macbook <macbook@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/12 05:27:59 by macbook           #+#    #+#             */
-/*   Updated: 2025/02/12 05:57:57 by macbook          ###   ########.fr       */
+/*   Updated: 2025/02/18 08:03:02 by macbook          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
